Move view decorators to project dir (minierp)
As all current decorators are common for both employee and hardware app,
we should move them to the project dir. This will keep better code placement
logic.

The views have to be updated to read from the new decorators location.
