+++ hardware/views.py
@@ -5,7 +5,7 @@ from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from django.core.paginator import Paginator
 
-from .decorators import  allowed_users
+from minierp.decorators import  allowed_users
 from .models import Laptop, Hardware
 from .forms import LaptopForm, LaptopReturnForm
 from .filters import LaptopFilter
